Merging with Master Fork

+++ src/Sylius/Component/Core/Uploader/ImageUploader.php
@@ -15,17 +15,14 @@ namespace Sylius\Component\Core\Uploader;
 
 use Gaufrette\Filesystem;
 use Sylius\Component\Core\Model\ImageInterface;
+use Symfony\Component\HttpFoundation\File\File;
+use Webmozart\Assert\Assert;
 
 class ImageUploader implements ImageUploaderInterface
 {
-    /**
-     * @var Filesystem
-     */
+    /** @var Filesystem */
     protected $filesystem;
 
-    /**
-     * @param Filesystem $filesystem
-     */
     public function __construct(Filesystem $filesystem)
     {
         $this->filesystem = $filesystem;
@@ -34,20 +31,25 @@ class ImageUploader implements ImageUploaderInterface
     /**
      * {@inheritdoc}
      */
-    public function upload(ImageInterface $image)
+    public function upload(ImageInterface $image): void
     {
         if (!$image->hasFile()) {
             return;
         }
+
+        $file = $image->getFile();
+
+        /** @var File $file */
+        Assert::isInstanceOf($file, File::class);
 
         if (null !== $image->getPath() && $this->has($image->getPath())) {
             $this->remove($image->getPath());
         }
 
         do {
-            $hash = md5(uniqid((string) mt_rand(), true));
-            $path = $this->expandPath($hash.'.'.$image->getFile()->guessExtension());
-        } while ($this->filesystem->has($path));
+            $hash = bin2hex(random_bytes(16));
+            $path = $this->expandPath($hash . '.' . $file->guessExtension());
+        } while ($this->isAdBlockingProne($path) || $this->filesystem->has($path));
 
         $image->setPath($path);
 
@@ -60,17 +62,16 @@ class ImageUploader implements ImageUploaderInterface
     /**
      * {@inheritdoc}
      */
-    public function remove($path)
+    public function remove(string $path): bool
     {
-        return $this->filesystem->delete($path);
+        if ($this->filesystem->has($path)) {
+            return $this->filesystem->delete($path);
+        }
+
+        return false;
     }
 
-    /**
-     * @param string $path
-     *
-     * @return string
-     */
-    private function expandPath($path)
+    private function expandPath(string $path): string
     {
         return sprintf(
             '%s/%s/%s',
@@ -80,13 +81,16 @@ class ImageUploader implements ImageUploaderInterface
         );
     }
 
-    /**
-     * @param string $path
-     *
-     * @return bool
-     */
-    private function has($path)
+    private function has(string $path): bool
     {
         return $this->filesystem->has($path);
+    }
+
+    /**
+     * Will return true if the path is prone to be blocked by ad blockers
+     */
+    private function isAdBlockingProne(string $path): bool
+    {
+        return strpos($path, 'ad') !== false;
     }
 }
